% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{Expression data}
\format{
An object of class \code{data.frame} with 200 rows and 30 columns.
}
\usage{
expression_data
}
\description{
Expression data
}
\keyword{datasets}
