% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcription_factors}
\alias{transcription_factors}
\title{Transcription factors}
\format{
An object of class \code{character} of length 1639.
}
\usage{
transcription_factors
}
\description{
Transcription factors
}
\keyword{datasets}
